# order-service/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Order Service",
        "version": "1.0.0"
    },
    "paths": {
        "/orders": {
            "get": {
                "summary": "Get all orders",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "summary": "Create an order",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "summary": "Get order by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Order": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "userId": {
                        "type": "string"
                    },
                    "bookId": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    }
                }
            }
        }
    }
}